Break .load() and .write() off into separate modules.
